// src/app/models/lego-set.model.ts
export interface LegoSet {
  engine: number;
  blue_short: number;
  blue_long: number;
  gray_45: number;
  gray_90: number;
  black_stick_short: number;
  black_short: number;
  red_shaft: number;
  gray_long: number;
  gray_short: number;
  green_3: number;
  gray_frame: number;
  white_45: number;
  wheel: number;
  tire: number;
  gray_loop: number;
}